emacs: Allow smarter counting of signature lines
Factor out the code that counts the lines of a signature, this allows
the function to easily be overridden if desired.

I have at the same time, added some basic code to detect inline PGP
markers that might occur after the signature, these now indicate the
end of the signature.