emacs: Allow all-tags searches to open in tree view
Modify notmuch-hello-query-counts so that the OPTIONS passed in can
contain an optional :search-type field.  If this exists then it is
copied into each of the returned searches that are created.

Then modify notmuch-hello-insert-alltags to pass ":search-type 'tree"
in its options list.  This is hard-coded for now, forcing the "All
Tags" view to open in tree mode.  This suits me, but will probably not
be acceptable for upstreaming.