emacs: Add optional arguments to notmuch-hello-insert-alltags
Update description string on notmuch-hello-sections to remove the text
saying that function in the sections list don't take arguments.  This
text was immediately contradicted in the next paragraph.   The new
text describes the current reality.

Add optional arguments to notmuch-hello-insert-alltags that are passed
through as options when generating the "All Tags" section.  This
allows (for example) for the :search-type to be specified.  This
change removes the hard-coding of the "All Tags" opening in tree mode
that the previous commit added.